add prunning test
