+++ cmd/nitro/init.go
@@ -10,93 +10,37 @@ import (
 	"fmt"
 	"math/big"
 	"os"
-	"reflect"
-	"regexp"
 	"runtime"
 	"strings"
 	"sync"
 	"time"
 
+	"github.com/offchainlabs/nitro/cmd/conf"
+	"github.com/offchainlabs/nitro/cmd/prunning"
 	"github.com/offchainlabs/nitro/cmd/util"
 
 	"github.com/cavaliergopher/grab/v3"
 	extract "github.com/codeclysm/extract/v3"
-	"github.com/ethereum/go-ethereum/accounts/abi/bind"
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/core"
 	"github.com/ethereum/go-ethereum/core/rawdb"
-	"github.com/ethereum/go-ethereum/core/state/pruner"
 	"github.com/ethereum/go-ethereum/core/types"
 	"github.com/ethereum/go-ethereum/ethdb"
 	"github.com/ethereum/go-ethereum/log"
 	"github.com/ethereum/go-ethereum/node"
 	"github.com/ethereum/go-ethereum/params"
-	"github.com/ethereum/go-ethereum/rpc"
 
 	"github.com/offchainlabs/nitro/arbnode"
-	"github.com/offchainlabs/nitro/arbnode/dataposter/storage"
 	"github.com/offchainlabs/nitro/arbos/arbosState"
 	"github.com/offchainlabs/nitro/arbos/arbostypes"
 	"github.com/offchainlabs/nitro/arbutil"
 	"github.com/offchainlabs/nitro/cmd/chaininfo"
 	"github.com/offchainlabs/nitro/cmd/ipfshelper"
 	"github.com/offchainlabs/nitro/execution/gethexec"
-	"github.com/offchainlabs/nitro/staker"
 	"github.com/offchainlabs/nitro/statetransfer"
-	"github.com/spf13/pflag"
 )
 
-type InitConfig struct {
-	Force           bool          `koanf:"force"`
-	Url             string        `koanf:"url"`
-	DownloadPath    string        `koanf:"download-path"`
-	DownloadPoll    time.Duration `koanf:"download-poll"`
-	DevInit         bool          `koanf:"dev-init"`
-	DevInitAddress  string        `koanf:"dev-init-address"`
-	DevInitBlockNum uint64        `koanf:"dev-init-blocknum"`
-	Empty           bool          `koanf:"empty"`
-	AccountsPerSync uint          `koanf:"accounts-per-sync"`
-	ImportFile      string        `koanf:"import-file"`
-	ThenQuit        bool          `koanf:"then-quit"`
-	Prune           string        `koanf:"prune"`
-	PruneBloomSize  uint64        `koanf:"prune-bloom-size"`
-	ResetToMessage  int64         `koanf:"reset-to-message"`
-}
-
-var InitConfigDefault = InitConfig{
-	Force:           false,
-	Url:             "",
-	DownloadPath:    "/tmp/",
-	DownloadPoll:    time.Minute,
-	DevInit:         false,
-	DevInitAddress:  "",
-	DevInitBlockNum: 0,
-	ImportFile:      "",
-	AccountsPerSync: 100000,
-	ThenQuit:        false,
-	Prune:           "",
-	PruneBloomSize:  2048,
-	ResetToMessage:  -1,
-}
-
-func InitConfigAddOptions(prefix string, f *pflag.FlagSet) {
-	f.Bool(prefix+".force", InitConfigDefault.Force, "if true: in case database exists init code will be reexecuted and genesis block compared to database")
-	f.String(prefix+".url", InitConfigDefault.Url, "url to download initializtion data - will poll if download fails")
-	f.String(prefix+".download-path", InitConfigDefault.DownloadPath, "path to save temp downloaded file")
-	f.Duration(prefix+".download-poll", InitConfigDefault.DownloadPoll, "how long to wait between polling attempts")
-	f.Bool(prefix+".dev-init", InitConfigDefault.DevInit, "init with dev data (1 account with balance) instead of file import")
-	f.String(prefix+".dev-init-address", InitConfigDefault.DevInitAddress, "Address of dev-account. Leave empty to use the dev-wallet.")
-	f.Uint64(prefix+".dev-init-blocknum", InitConfigDefault.DevInitBlockNum, "Number of preinit blocks. Must exist in ancient database.")
-	f.Bool(prefix+".empty", InitConfigDefault.Empty, "init with empty state")
-	f.Bool(prefix+".then-quit", InitConfigDefault.ThenQuit, "quit after init is done")
-	f.String(prefix+".import-file", InitConfigDefault.ImportFile, "path for json data to import")
-	f.Uint(prefix+".accounts-per-sync", InitConfigDefault.AccountsPerSync, "during init - sync database every X accounts. Lower value for low-memory systems. 0 disables.")
-	f.String(prefix+".prune", InitConfigDefault.Prune, "pruning for a given use: \"full\" for full nodes serving RPC requests, or \"validator\" for validators")
-	f.Uint64(prefix+".prune-bloom-size", InitConfigDefault.PruneBloomSize, "the amount of memory in megabytes to use for the pruning bloom filter (higher values prune better)")
-	f.Int64(prefix+".reset-to-message", InitConfigDefault.ResetToMessage, "forces a reset to an old message height. Also set max-reorg-resequence-depth=0 to force re-reading messages")
-}
-
-func downloadInit(ctx context.Context, initConfig *InitConfig) (string, error) {
+func downloadInit(ctx context.Context, initConfig *conf.InitConfig) (string, error) {
 	if initConfig.Url == "" {
 		return "", nil
 	}
@@ -215,228 +159,6 @@ func validateBlockChain(blockChain *core.BlockChain, chainConfig *params.ChainCo
 	return nil
 }
 
-type importantRoots struct {
-	chainDb ethdb.Database
-	roots   []common.Hash
-	heights []uint64
-}
-
-// The minimum block distance between two important roots
-const minRootDistance = 2000
-
-// Marks a header as important, and records its root and height.
-// If overwrite is true, it'll remove any future roots and replace them with this header.
-// If overwrite is false, it'll ignore this header if it has future roots.
-func (r *importantRoots) addHeader(header *types.Header, overwrite bool) error {
-	targetBlockNum := header.Number.Uint64()
-	for {
-		if header == nil || header.Root == (common.Hash{}) {
-			log.Error("missing state of pruning target", "blockNum", targetBlockNum)
-			return nil
-		}
-		exists, err := r.chainDb.Has(header.Root.Bytes())
-		if err != nil {
-			return err
-		}
-		if exists {
-			break
-		}
-		num := header.Number.Uint64()
-		if num%3000 == 0 {
-			log.Info("looking for old block with state to keep", "current", num, "target", targetBlockNum)
-		}
-		// An underflow is fine here because it'll just return nil due to not found
-		header = rawdb.ReadHeader(r.chainDb, header.ParentHash, num-1)
-	}
-	height := header.Number.Uint64()
-	for len(r.heights) > 0 && r.heights[len(r.heights)-1] > height {
-		if !overwrite {
-			return nil
-		}
-		r.roots = r.roots[:len(r.roots)-1]
-		r.heights = r.heights[:len(r.heights)-1]
-	}
-	if len(r.heights) > 0 && r.heights[len(r.heights)-1]+minRootDistance > height {
-		return nil
-	}
-	r.roots = append(r.roots, header.Root)
-	r.heights = append(r.heights, height)
-	return nil
-}
-
-var hashListRegex = regexp.MustCompile("^(0x)?[0-9a-fA-F]{64}(,(0x)?[0-9a-fA-F]{64})*$")
-
-// Finds important roots to retain while proving
-func findImportantRoots(ctx context.Context, chainDb ethdb.Database, stack *node.Node, nodeConfig *NodeConfig, cacheConfig *core.CacheConfig, l1Client arbutil.L1Interface, rollupAddrs chaininfo.RollupAddresses) ([]common.Hash, error) {
-	initConfig := &nodeConfig.Init
-	chainConfig := gethexec.TryReadStoredChainConfig(chainDb)
-	if chainConfig == nil {
-		return nil, errors.New("database doesn't have a chain config (was this node initialized?)")
-	}
-	arbDb, err := stack.OpenDatabase("arbitrumdata", 0, 0, "", true)
-	if err != nil {
-		return nil, err
-	}
-	defer func() {
-		err := arbDb.Close()
-		if err != nil {
-			log.Warn("failed to close arbitrum database after finding pruning targets", "err", err)
-		}
-	}()
-	roots := importantRoots{
-		chainDb: chainDb,
-	}
-	genesisNum := chainConfig.ArbitrumChainParams.GenesisBlockNum
-	genesisHash := rawdb.ReadCanonicalHash(chainDb, genesisNum)
-	genesisHeader := rawdb.ReadHeader(chainDb, genesisHash, genesisNum)
-	if genesisHeader == nil {
-		return nil, errors.New("missing L2 genesis block header")
-	}
-	err = roots.addHeader(genesisHeader, false)
-	if err != nil {
-		return nil, err
-	}
-	if initConfig.Prune == "validator" {
-		if l1Client == nil || reflect.ValueOf(l1Client).IsNil() {
-			return nil, errors.New("an L1 connection is required for validator pruning")
-		}
-		callOpts := bind.CallOpts{
-			Context:     ctx,
-			BlockNumber: big.NewInt(int64(rpc.FinalizedBlockNumber)),
-		}
-		rollup, err := staker.NewRollupWatcher(rollupAddrs.Rollup, l1Client, callOpts)
-		if err != nil {
-			return nil, err
-		}
-		latestConfirmedNum, err := rollup.LatestConfirmed(&callOpts)
-		if err != nil {
-			return nil, err
-		}
-		latestConfirmedNode, err := rollup.LookupNode(ctx, latestConfirmedNum)
-		if err != nil {
-			return nil, err
-		}
-		confirmedHash := latestConfirmedNode.Assertion.AfterState.GlobalState.BlockHash
-		confirmedNumber := rawdb.ReadHeaderNumber(chainDb, confirmedHash)
-		var confirmedHeader *types.Header
-		if confirmedNumber != nil {
-			confirmedHeader = rawdb.ReadHeader(chainDb, confirmedHash, *confirmedNumber)
-		}
-		if confirmedHeader != nil {
-			err = roots.addHeader(confirmedHeader, false)
-			if err != nil {
-				return nil, err
-			}
-		} else {
-			log.Warn("missing latest confirmed block", "hash", confirmedHash)
-		}
-
-		validatorDb := rawdb.NewTable(arbDb, storage.BlockValidatorPrefix)
-		lastValidated, err := staker.ReadLastValidatedInfo(validatorDb)
-		if err != nil {
-			return nil, err
-		}
-		if lastValidated != nil {
-			var lastValidatedHeader *types.Header
-			headerNum := rawdb.ReadHeaderNumber(chainDb, lastValidated.GlobalState.BlockHash)
-			if headerNum != nil {
-				lastValidatedHeader = rawdb.ReadHeader(chainDb, lastValidated.GlobalState.BlockHash, *headerNum)
-			}
-			if lastValidatedHeader != nil {
-				err = roots.addHeader(lastValidatedHeader, false)
-				if err != nil {
-					return nil, err
-				}
-			} else {
-				log.Warn("missing latest validated block", "hash", lastValidated.GlobalState.BlockHash)
-			}
-		}
-	} else if initConfig.Prune == "full" {
-		if nodeConfig.Node.ValidatorRequired() {
-			return nil, errors.New("refusing to prune to full-node level when validator is enabled (you should prune in validator mode)")
-		}
-	} else if hashListRegex.MatchString(initConfig.Prune) {
-		parts := strings.Split(initConfig.Prune, ",")
-		roots := []common.Hash{genesisHeader.Root}
-		for _, part := range parts {
-			root := common.HexToHash(part)
-			if root == genesisHeader.Root {
-				// This was already included in the builtin list
-				continue
-			}
-			roots = append(roots, root)
-		}
-		return roots, nil
-	} else {
-		return nil, fmt.Errorf("unknown pruning mode: \"%v\"", initConfig.Prune)
-	}
-	if l1Client != nil {
-		// Find the latest finalized block and add it as a pruning target
-		l1Block, err := l1Client.BlockByNumber(ctx, big.NewInt(int64(rpc.FinalizedBlockNumber)))
-		if err != nil {
-			return nil, fmt.Errorf("failed to get finalized block: %w", err)
-		}
-		l1BlockNum := l1Block.NumberU64()
-		tracker, err := arbnode.NewInboxTracker(arbDb, nil, nil)
-		if err != nil {
-			return nil, err
-		}
-		batch, err := tracker.GetBatchCount()
-		if err != nil {
-			return nil, err
-		}
-		for {
-			if ctx.Err() != nil {
-				return nil, ctx.Err()
-			}
-			if batch == 0 {
-				// No batch has been finalized
-				break
-			}
-			batch -= 1
-			meta, err := tracker.GetBatchMetadata(batch)
-			if err != nil {
-				return nil, err
-			}
-			if meta.ParentChainBlock <= l1BlockNum {
-				signedBlockNum := arbutil.MessageCountToBlockNumber(meta.MessageCount, genesisNum)
-				blockNum := uint64(signedBlockNum)
-				l2Hash := rawdb.ReadCanonicalHash(chainDb, blockNum)
-				l2Header := rawdb.ReadHeader(chainDb, l2Hash, blockNum)
-				if l2Header == nil {
-					log.Warn("latest finalized L2 block is unknown", "blockNum", signedBlockNum)
-					break
-				}
-				err = roots.addHeader(l2Header, false)
-				if err != nil {
-					return nil, err
-				}
-				break
-			}
-		}
-	}
-	roots.roots = append(roots.roots, common.Hash{}) // the latest snapshot
-	log.Info("found pruning target blocks", "heights", roots.heights, "roots", roots.roots)
-	return roots.roots, nil
-}
-
-func pruneChainDb(ctx context.Context, chainDb ethdb.Database, stack *node.Node, nodeConfig *NodeConfig, cacheConfig *core.CacheConfig, l1Client arbutil.L1Interface, rollupAddrs chaininfo.RollupAddresses) error {
-	config := &nodeConfig.Init
-	if config.Prune == "" {
-		return pruner.RecoverPruning(stack.InstanceDir(), chainDb)
-	}
-	root, err := findImportantRoots(ctx, chainDb, stack, nodeConfig, cacheConfig, l1Client, rollupAddrs)
-	if err != nil {
-		return fmt.Errorf("failed to find root to retain for pruning: %w", err)
-	}
-
-	pruner, err := pruner.NewPruner(chainDb, pruner.Config{Datadir: stack.InstanceDir(), BloomSize: config.PruneBloomSize})
-	if err != nil {
-		return err
-	}
-	return pruner.Prune(root)
-}
-
 func openInitializeChainDb(ctx context.Context, stack *node.Node, config *NodeConfig, chainId *big.Int, cacheConfig *core.CacheConfig, l1Client arbutil.L1Interface, rollupAddrs chaininfo.RollupAddresses) (ethdb.Database, *core.BlockChain, error) {
 	if !config.Init.Force {
 		if readOnlyDb, err := stack.OpenDatabaseWithFreezer("l2chaindata", 0, 0, "", "", true); err == nil {
@@ -446,7 +168,7 @@ func openInitializeChainDb(ctx context.Context, stack *node.Node, config *NodeCo
 				if err != nil {
 					return chainDb, nil, err
 				}
-				err = pruneChainDb(ctx, chainDb, stack, config, cacheConfig, l1Client, rollupAddrs)
+				err = prunning.PruneChainDb(ctx, chainDb, stack, &config.Init, cacheConfig, l1Client, rollupAddrs, config.Node.ValidatorRequired())
 				if err != nil {
 					return chainDb, nil, fmt.Errorf("error pruning: %w", err)
 				}
@@ -642,7 +364,7 @@ func openInitializeChainDb(ctx context.Context, stack *node.Node, config *NodeCo
 		return chainDb, l2BlockChain, err
 	}
 
-	err = pruneChainDb(ctx, chainDb, stack, config, cacheConfig, l1Client, rollupAddrs)
+	err = prunning.PruneChainDb(ctx, chainDb, stack, &config.Init, cacheConfig, l1Client, rollupAddrs, config.Node.ValidatorRequired())
 	if err != nil {
 		return chainDb, nil, fmt.Errorf("error pruning: %w", err)
 	}
